+++ src/snowflake/cli/api/project/schemas/native_app/path_mapping.py
@@ -6,9 +6,8 @@ from pydantic import Field
 from snowflake.cli.api.project.schemas.updatable_model import UpdatableModel
 
 # Limited to python options only for now.
-ProcessorOptions = Literal["python-snowpark", "PYTHON-SNOWPARK"]
-VirtualEnvTypeOptions = Literal["conda", "CONDA", "venv", "VENV"]
-LanguageOptions = Literal["python", "PYTHON"]
+# Also, processor name can only be "snowpark" for now.
+VirtualEnvTypeOptions = Literal["conda", "venv"]
 
 
 class VirtualEnvironment(UpdatableModel):
@@ -22,17 +21,11 @@ class VirtualEnvironment(UpdatableModel):
 
 
 class AnnotationProcessor(UpdatableModel):
-    language: LanguageOptions = Field(
-        title="Programming language in which the annotated code is written.",
-    )
-    language_version: str = Field(
-        title="Programming language version in which the annotated code is written, and to use in a virtual environment, if provided.",
-    )
-    name: ProcessorOptions = Field(
+    name: str = Field(
         title="Name of the processor to be invoked to discover the annotated code."
     )
     virtual_env: Optional[VirtualEnvironment] = Field(
-        title="The virtual environment to run the processor in.",
+        title="The virtual environment to run the processor in. If none is provided, then the Snowflake CLI will try to get the current environment information.",
         default=None,
     )
 
@@ -40,4 +33,4 @@ class AnnotationProcessor(UpdatableModel):
 class PathMapping(UpdatableModel):
     src: str
     dest: Optional[str] = None
-    processors: Optional[List[AnnotationProcessor]] = None
+    processors: Optional[List[str | AnnotationProcessor]] = None
