--- tests/test_data/projects/napp_with_annotation_processor_failure/snowflake.yml
@@ -1,25 +1,0 @@
-definition_version: 1
-native_app:
-  name: myapp
-
-  # the following shows how to properly use a quoted identifier in YAML
-  source_stage: >-
-    "MySourceSchema"."SRC_Stage"
-
-  artifacts:
-    - setup.sql
-    - app/README.md
-    - src: app/streamlit/*.py
-      dest: ui/
-      processors:
-        - name: python-snowpark
-          language: python
-          virtual_env:
-            name: snowpark_venv
-            env_type: venv
-          
-
-  package:
-    scripts:
-      - 001-shared.sql
-      - 002-shared.sql
